Change float to double

+++ src/core/include/cesium/omniverse/OmniImagery.h
@@ -13,7 +13,7 @@ class OmniImagery {
     [[nodiscard]] int64_t getIonAssetId() const;
     [[nodiscard]] std::optional<CesiumIonClient::Token> getIonAccessToken() const;
     [[nodiscard]] bool getShowCreditsOnScreen() const;
-    [[nodiscard]] float getAlpha() const;
+    [[nodiscard]] double getAlpha() const;
 
   private:
     pxr::SdfPath _path;
